uvc: avoid using systemd script manipulation
When this app is run in oakapp container the systemd init scripts are not exposed to the container, therefore we can't
modify them.

In OS 1.18.3-uvc-v6 and onwards we support dynamic UDC binding/unbinding without manipulation of adb deamon. This
is how we can simply bind/unbind the UDC with echo commands

Signed-off-by: Jure Rebernik <jure.rebernik@luxonis.com>